requirements.txt
argcomplete==1.9.3
awacs==0.8.2
boto3==1.9.97
botocore==1.12.97
cachetools==3.1.0
certifi==2018.11.29
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
docutils==0.14
durationpy==0.5
Flask==1.0.2
Flask-Cors==3.0.7
future==0.16.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-language==1.1.1
googleapis-common-protos==1.5.8
grpcio==1.18.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.0
nltk==3.4
oauthlib==3.0.1
placebo==0.9.0
protobuf==3.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
PySocks==1.6.8
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
textblob==0.15.2
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.4
tweepy==3.7.0
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.1
